rename controller to ui package
