+++ src/main/java/seedu/address/ui/MainWindow.java
@@ -1,4 +1,4 @@
-package seedu.address.controller;
+package seedu.address.ui;
 
 import javafx.fxml.FXML;
 import javafx.scene.Node;
