+++ src/main/java/seedu/address/ui/UiPartLoader.java
@@ -1,4 +1,4 @@
-package seedu.address.controller;
+package seedu.address.ui;
 
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Node;
